requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
hypothesis>=6.92.0
pytest>=7.4.0
google-generativeai>=0.3.0
pandas>=2.0.0
ipython>=8.0.0
